# docs/swagger.yaml
definitions:
...
  vo.Pagination:
    properties:
      curPage:
        type: integer
      items: {}
      totalNum:
        type: integer
    type: object
...
  vo.ReadMessage:
    properties:
      id:
        type: integer
    type: object
  vo.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  vo.UserMessage:
    properties:
      content:
        type: string
      id:
        type: integer
      state:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /message/my:
    get:
      consumes:
      - application/json
      description: get my message
      parameters:
      - description: language
        in: header
        name: lang
        required: true
        type: string
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: state 1：unRead 2:read
        in: query
        name: state
        type: integer
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.UserMessage'
                        type: array
                    type: object
              type: object
      summary: GetMyMessages
      tags:
      - message
  /message/read:
    post:
      consumes:
      - application/json
      description: make message read
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.ReadMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: ReadMessage
      tags:
      - message
...
swagger: "2.0"